fixup! treewide: convert enums to ints before formatting

+++ trace2/tr2_sysenv.c
@@ -92,7 +92,7 @@ void tr2_sysenv_load(void)
 const char *tr2_sysenv_get(enum tr2_sysenv_variable var)
 {
 	if (var >= TR2_SYSENV_MUST_BE_LAST)
-		BUG("tr2_sysenv_get invalid var '%d'", var);
+		BUG("tr2_sysenv_get invalid var '%d'", (int)var);
 
 	if (!tr2_sysenv_settings[var].getenv_called) {
 		const char *v = getenv(tr2_sysenv_settings[var].env_var_name);
